% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDT.R
\name{get_shaka_samvat}
\alias{get_shaka_samvat}
\title{Used to calculate the Shaka samvat year from system date}
\usage{
get_shaka_samvat()
}
\description{
Used to calculate the Shaka samvat year from system date
}
