% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDT.R
\name{samvatsar}
\alias{samvatsar}
\title{Used to retrieve the samvatsar}
\usage{
samvatsar(year, month = 4)
}
\arguments{
\item{year}{Year as an integer}

\item{month}{Month as an integer}
}
\description{
Used to retrieve the samvatsar
}
