% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDT.R
\docType{data}
\name{samvatsar_list}
\alias{samvatsar_list}
\title{VedicDateTime
Prajwal Patil
A list containing all the Samvatsars in chronological order}
\format{
An object of class \code{list} of length 60.
}
\usage{
samvatsar_list
}
\description{
VedicDateTime
Prajwal Patil
A list containing all the Samvatsars in chronological order
}
\keyword{datasets}
