% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDT.R
\name{shaka_samvat}
\alias{shaka_samvat}
\title{Used to calculate the Shaka samvat year from a given year and month}
\usage{
shaka_samvat(year, month = 4)
}
\arguments{
\item{year}{Year as an integer}

\item{month}{Month as an integer}
}
\description{
Used to calculate the Shaka samvat year from a given year and month
}
